GitHub Actions Fix URL and Version Prefix in Generate Client Workflow
This commit addresses an issue in the Generate Client workflow where an incorrect resource URL was being requested. The workflow has been updated to ensure that the URL includes the correct version, taking into account the presence or absence of the "v" prefix. Additionally, a new step has been added to check the version prefix and update it if necessary, guaranteeing that the correct version is used throughout the workflow. These changes improve the reliability and accuracy of the client generation process.